move News item
