--- src/components/App/index.jsx
@@ -1,16 +1,0 @@
-function App() {
-  return (
-    <div className="App">
-      <a
-        className="App-link"
-        href="https://reactjs.org"
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        Learn React
-      </a>
-    </div>
-  );
-}
-
-export default App;
